Refactor code to remove the need for the fdb client lib
