+++ controllers/admin_client.go
@@ -27,6 +27,8 @@ import (
 	"strings"
 	"sync"
 
+	"github.com/FoundationDB/fdb-kubernetes-operator/internal"
+
 	fdbtypes "github.com/FoundationDB/fdb-kubernetes-operator/api/v1beta1"
 	corev1 "k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -248,7 +250,7 @@ func (client *MockAdminClient) GetStatus() (*fdbtypes.FoundationDBStatus, error)
 
 			status.Cluster.Processes[fmt.Sprintf("%s-%d", pod.Name, processIndex)] = fdbtypes.FoundationDBStatusProcessInfo{
 				Address:       fullAddress,
-				ProcessClass:  GetProcessClassFromMeta(pod.ObjectMeta),
+				ProcessClass:  internal.GetProcessClassFromMeta(pod.ObjectMeta),
 				CommandLine:   command,
 				Excluded:      excluded,
 				Locality:      locality,
