+++ controllers/pod_models_test.go
@@ -23,9 +23,9 @@ package controllers
 import (
 	"fmt"
 	"reflect"
-	"strings"
 
 	fdbtypes "github.com/FoundationDB/fdb-kubernetes-operator/api/v1beta1"
+	"github.com/FoundationDB/fdb-kubernetes-operator/internal"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
@@ -42,7 +42,7 @@ var _ = Describe("pod_models", func() {
 
 	BeforeEach(func() {
 		cluster = createDefaultCluster()
-		err := NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+		err := internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 		Expect(err).NotTo(HaveOccurred())
 	})
 
@@ -126,7 +126,7 @@ var _ = Describe("pod_models", func() {
 			})
 
 			It("should add the annotations to the metadata", func() {
-				hash, err := GetPodSpecHash(cluster, processClassFromLabels(pod.Labels), 1, &pod.Spec)
+				hash, err := GetPodSpecHash(cluster, internal.ProcessClassFromLabels(pod.Labels), 1, &pod.Spec)
 				Expect(err).NotTo(HaveOccurred())
 				Expect(pod.ObjectMeta.Annotations).To(Equal(map[string]string{
 					"fdb-annotation":                     "value1",
@@ -146,7 +146,7 @@ var _ = Describe("pod_models", func() {
 						},
 					},
 				}}}
-				err := NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err := internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				pod, err = GetPod(cluster, fdbtypes.ProcessClassStorage, 1)
@@ -738,7 +738,7 @@ var _ = Describe("pod_models", func() {
 						},
 					},
 				}}}
-				err := NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err := internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
@@ -766,7 +766,7 @@ var _ = Describe("pod_models", func() {
 						},
 					},
 				}}}
-				err := NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err := internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
@@ -920,7 +920,7 @@ var _ = Describe("pod_models", func() {
 						}},
 					},
 				}}}
-				err = NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err = internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
@@ -976,7 +976,7 @@ var _ = Describe("pod_models", func() {
 						}},
 					},
 				}}}
-				err = NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err = internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 			})
 
@@ -1246,7 +1246,7 @@ var _ = Describe("pod_models", func() {
 						},
 					},
 				}}}
-				err = NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err = internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
@@ -1580,7 +1580,7 @@ var _ = Describe("pod_models", func() {
 		Context("with custom pvc", func() {
 			BeforeEach(func() {
 				cluster.Spec.Processes = map[fdbtypes.ProcessClass]fdbtypes.ProcessSettings{fdbtypes.ProcessClassGeneral: {VolumeClaimTemplate: &corev1.PersistentVolumeClaim{ObjectMeta: metav1.ObjectMeta{Name: "claim1"}}}}
-				err = NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err = internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
@@ -1702,7 +1702,7 @@ var _ = Describe("pod_models", func() {
 						},
 					},
 				}
-				err = NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err = internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).ToNot(HaveOccurred())
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
 				Expect(err).ToNot(HaveOccurred())
@@ -1737,7 +1737,7 @@ var _ = Describe("pod_models", func() {
 						},
 					},
 				}
-				err = NormalizeClusterSpec(&cluster.Spec, DeprecationOptions{})
+				err = internal.NormalizeClusterSpec(&cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).ToNot(HaveOccurred())
 				spec, err = GetPodSpec(cluster, fdbtypes.ProcessClassStorage, 1)
 				Expect(err).ToNot(HaveOccurred())
@@ -2392,7 +2392,7 @@ var _ = Describe("pod_models", func() {
 
 		Describe("deprecations", func() {
 			JustBeforeEach(func() {
-				err := NormalizeClusterSpec(spec, DeprecationOptions{})
+				err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 			})
 
@@ -2733,7 +2733,7 @@ var _ = Describe("pod_models", func() {
 						"knob_disable_posix_kernel_aio = 1",
 						"knob_disable_posix_kernel_aio = 1",
 					}
-					err := NormalizeClusterSpec(spec, DeprecationOptions{})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{})
 					Expect(err).To(HaveOccurred())
 				})
 			})
@@ -2743,7 +2743,7 @@ var _ = Describe("pod_models", func() {
 					spec.CustomParameters = []string{
 						"datadir=1",
 					}
-					err := NormalizeClusterSpec(spec, DeprecationOptions{})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{})
 					Expect(err).To(HaveOccurred())
 				})
 			})
@@ -2758,7 +2758,7 @@ var _ = Describe("pod_models", func() {
 							},
 						},
 					}
-					err := NormalizeClusterSpec(spec, DeprecationOptions{})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{})
 					Expect(err).To(HaveOccurred())
 				})
 			})
@@ -2772,7 +2772,7 @@ var _ = Describe("pod_models", func() {
 							},
 						},
 					}
-					err := NormalizeClusterSpec(spec, DeprecationOptions{})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{})
 					Expect(err).To(HaveOccurred())
 				})
 			})
@@ -2781,7 +2781,7 @@ var _ = Describe("pod_models", func() {
 		Describe("defaults", func() {
 			Context("with the current defaults", func() {
 				JustBeforeEach(func() {
-					err := NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: false})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: false})
 					Expect(err).NotTo(HaveOccurred())
 				})
 
@@ -2919,7 +2919,7 @@ var _ = Describe("pod_models", func() {
 
 			Context("with the current defaults, changes only", func() {
 				JustBeforeEach(func() {
-					err := NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: true})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: true})
 					Expect(err).NotTo(HaveOccurred())
 				})
 
@@ -2956,7 +2956,7 @@ var _ = Describe("pod_models", func() {
 
 			Context("with the future defaults", func() {
 				JustBeforeEach(func() {
-					err := NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: true, OnlyShowChanges: false})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: true, OnlyShowChanges: false})
 					Expect(err).NotTo(HaveOccurred())
 				})
 
@@ -3094,7 +3094,7 @@ var _ = Describe("pod_models", func() {
 
 			Context("with the future defaults, changes only", func() {
 				JustBeforeEach(func() {
-					err := NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: true, OnlyShowChanges: true})
+					err := internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: true, OnlyShowChanges: true})
 					Expect(err).NotTo(HaveOccurred())
 				})
 
@@ -3123,13 +3123,13 @@ var _ = Describe("pod_models", func() {
 				var originalSpec *fdbtypes.FoundationDBClusterSpec
 
 				BeforeEach(func() {
-					err = NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: true})
+					err = internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: true})
 					Expect(err).NotTo(HaveOccurred())
 					originalSpec = spec.DeepCopy()
 				})
 
 				JustBeforeEach(func() {
-					err = NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: true, OnlyShowChanges: true})
+					err = internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: true, OnlyShowChanges: true})
 					Expect(err).NotTo(HaveOccurred())
 				})
 
@@ -3166,7 +3166,7 @@ var _ = Describe("pod_models", func() {
 				processSetting.AllowTagOverride = &allowTagOverride
 				cluster.Spec.Processes[fdbtypes.ProcessClassGeneral] = processSetting
 
-				err = NormalizeClusterSpec(spec, DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: true})
+				err = internal.NormalizeClusterSpec(spec, internal.DeprecationOptions{UseFutureDefaults: false, OnlyShowChanges: true})
 				Expect(err).NotTo(HaveOccurred())
 			})
 			It("should set the image name with tag", func() {
@@ -3291,48 +3291,5 @@ var _ = Describe("pod_models", func() {
 					}, "myimage:mytag"),
 			)
 		})
-	})
-
-	When("Providing a custom parameter", func() {
-		type testCase struct {
-			Input              []string
-			ExpectedViolations []string
-		}
-
-		DescribeTable("should set the expected parameter or print the violation",
-			func(tc testCase) {
-				err := ValidateCustomParameters(tc.Input)
-				var errMsg error
-				if len(tc.ExpectedViolations) > 0 {
-					errMsg = fmt.Errorf("found the following customParameters violations:\n%s", strings.Join(tc.ExpectedViolations, "\n"))
-				}
-
-				if err == nil {
-					Expect(len(tc.ExpectedViolations)).To(BeNumerically("==", 0))
-				} else {
-					Expect(err).To(Equal(errMsg))
-				}
-			},
-			Entry("Valid parameter without duplicate",
-				testCase{
-					Input:              []string{"blahblah=1"},
-					ExpectedViolations: []string{},
-				}),
-			Entry("Valid parameter with duplicate",
-				testCase{
-					Input:              []string{"blahblah=1", "blahblah=1"},
-					ExpectedViolations: []string{"found duplicated customParameter: blahblah"},
-				}),
-			Entry("Protected parameter without duplicate",
-				testCase{
-					Input:              []string{"datadir=1"},
-					ExpectedViolations: []string{"found protected customParameter: datadir, please remove this parameter from the customParameters list"},
-				}),
-			Entry("Valid parameter with duplicate and protected parameter",
-				testCase{
-					Input:              []string{"blahblah=1", "blahblah=1", "datadir=1"},
-					ExpectedViolations: []string{"found duplicated customParameter: blahblah", "found protected customParameter: datadir, please remove this parameter from the customParameters list"},
-				}),
-		)
 	})
 })
