+++ controllers/update_sidecar_versions_test.go
@@ -22,6 +22,7 @@ package controllers
 
 import (
 	fdbtypes "github.com/FoundationDB/fdb-kubernetes-operator/api/v1beta1"
+	"github.com/FoundationDB/fdb-kubernetes-operator/internal"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
@@ -49,7 +50,7 @@ var _ = Describe("update_sidecar_versions", func() {
 
 		DescribeTable("should return the correct image",
 			func(input testCase, expected string) {
-				err := NormalizeClusterSpec(&input.cluster.Spec, DeprecationOptions{})
+				err := internal.NormalizeClusterSpec(&input.cluster.Spec, internal.DeprecationOptions{})
 				Expect(err).NotTo(HaveOccurred())
 
 				image, err := getSidecarImage(input.cluster, input.instance)
